+++ src/components/topbar/TopBar.tsx
@@ -1,12 +1,23 @@
 import React from 'react';
+import {
+	Button as ReactAriaButton,
+	DialogTrigger as ReactAriaDialogTrigger,
+	Popover as ReactAriaPopover,
+} from 'react-aria-components';
+import { AvatarButton } from '../../avatar-button';
+import { AvatarLink } from '../../avatar-link';
 import type { TopBarToolNameProps } from '../../TopBar';
 import { mergeDeep } from '../../util/mergeDeep';
+import { useResize } from '../../util/useResize';
 import { Avatar } from '../avatar/Avatar';
+import { Icon } from '../icon/Icon';
 import type { TopBarTheme } from './styles';
 import {
 	defaultTopBarTheme,
-	topBarContainerLeftStyles,
-	topBarContainerRightStyles,
+	topBarCollapsedNavMenuButtonStyles,
+	topBarCollapsedNavMenuPopoverStyles,
+	topBarContainerStyles,
+	topBarSpacerStyles,
 	topBarStyles,
 } from './styles';
 import { TopBarItem } from './topBarItem/TopBarItem';
@@ -86,17 +97,23 @@ export function TopBarContainerLeft({
 
 export function TopBar({
 	children,
+	collapseBelow = {
+		toolName: 'lg',
+		containerLeft: 'lg',
+	},
 	theme = {},
 	cssOverrides,
 	className,
 	...props
 }: TopBarProps) {
+	const [menuOpen, setMenuOpen] = React.useState(false);
+	const buttonRef = React.useRef<HTMLButtonElement>(null);
 	const mergedTheme = mergeDeep(defaultTopBarTheme, theme);
 
-	let leftSide: React.ReactElement | null = null;
-	let rightSide: React.ReactElement | null = null;
-	let toolName: React.ReactElement | null = null;
-	let avatar: React.ReactElement | null = null;
+	let leftSide: React.ReactElement | null | undefined;
+	let rightSide: React.ReactElement | null | undefined;
+	let toolName: React.ReactElement | null | undefined;
+	let avatar: React.ReactElement | null | undefined;
 
 	React.Children.forEach(children, (child) => {
 		if (!React.isValidElement(child)) {
@@ -109,14 +126,18 @@ export function TopBar({
 		if (child.type === TopBarToolName) {
 			toolName ??= React.cloneElement(
 				child as React.ReactElement<TopBarToolNameProps>,
-				{ theme: mergedTheme.ToolName },
+				{ theme: mergedTheme.ToolName, collapseBelow: collapseBelow.toolName },
 			);
 		}
 
 		/**
 		 * Accepts an avatar that will always be rendered on the right hand side, within an item for styling
 		 */
-		if (child.type === Avatar) {
+		if (
+			child.type === Avatar ||
+			child.type === AvatarLink ||
+			child.type === AvatarButton
+		) {
 			avatar ??= (
 				<TopBarItem theme={mergedTheme.Item} alignment="right">
 					{child}
@@ -141,6 +162,49 @@ export function TopBar({
 		}
 	});
 
+	const leftSideMenuItems: React.ReactElement[] = [];
+
+	if (leftSide) {
+		React.Children.forEach(
+			(leftSide as React.ReactElement<{ children?: React.ReactNode }>).props
+				.children as React.ReactElement[],
+			(child) => {
+				if (!React.isValidElement(child)) {
+					return;
+				}
+
+				if (child.type === TopBarNavigation) {
+					leftSideMenuItems.push(
+						React.cloneElement(
+							child as React.ReactElement<TopBarNavigationProps>,
+							{
+								theme: mergedTheme.Navigation,
+								alignment: 'left',
+								_menuOpen: menuOpen,
+							},
+						),
+					);
+				}
+
+				if (child.type === TopBarItem) {
+					leftSideMenuItems.push(
+						React.cloneElement(child as React.ReactElement<TopBarItemProps>, {
+							theme: mergedTheme.Item,
+							alignment: 'left',
+							_menuOpen: menuOpen,
+						}),
+					);
+				}
+			},
+		);
+	}
+
+	useResize(() => {
+		if (menuOpen) {
+			setMenuOpen(false);
+		}
+	});
+
 	return (
 		<div
 			css={[topBarStyles(mergedTheme), cssOverrides]}
@@ -148,15 +212,68 @@ export function TopBar({
 			{...props}
 		>
 			{/* LHS */}
-			<div css={topBarContainerLeftStyles(mergedTheme)}>
-				{toolName}
+			{leftSideMenuItems.length > 0 && (
+				<div
+					css={topBarContainerStyles(mergedTheme, {
+						showUntil: collapseBelow.containerLeft,
+					})}
+				>
+					<ReactAriaDialogTrigger
+						isOpen={menuOpen}
+						onOpenChange={(isOpen) => {
+							setMenuOpen(isOpen);
+						}}
+					>
+						<ReactAriaButton
+							aria-label="Navigation Menu"
+							ref={buttonRef}
+							css={topBarCollapsedNavMenuButtonStyles(mergedTheme, menuOpen)}
+						>
+							<TopBarItem
+								alignment="left"
+								size="sm"
+								theme={mergedTheme.Item}
+								aria-label="Navigation Menu"
+							>
+								<Icon
+									fill={mergedTheme.collapsedNavMenu.button.color}
+									size="lg"
+								>
+									{menuOpen ? 'close' : 'menu'}
+								</Icon>
+							</TopBarItem>
+						</ReactAriaButton>
+						<ReactAriaPopover
+							css={topBarCollapsedNavMenuPopoverStyles(mergedTheme)}
+							offset={mergedTheme.collapsedNavMenu.popover.offset}
+							containerPadding={
+								mergedTheme.collapsedNavMenu.popover.containerPadding
+							}
+							shouldFlip={false}
+						>
+							{leftSideMenuItems}
+						</ReactAriaPopover>
+					</ReactAriaDialogTrigger>
+				</div>
+			)}
+			<div css={topBarContainerStyles(mergedTheme)}>{toolName}</div>
+			<div
+				css={topBarContainerStyles(mergedTheme, {
+					collapseBelow: collapseBelow.containerLeft,
+				})}
+			>
 				{leftSide}
 			</div>
-			{/* RHS */}
-			<div css={topBarContainerRightStyles(mergedTheme)}>
+			{/* RHS - topBarSpacerStyles pushes content to the right */}
+			<div
+				css={[
+					topBarSpacerStyles(mergedTheme),
+					topBarContainerStyles(mergedTheme),
+				]}
+			>
 				{rightSide}
-				{avatar}
 			</div>
+			<div css={topBarContainerStyles(mergedTheme)}>{avatar}</div>
 		</div>
 	);
 }