+++ src/components/topbar/styles.ts
@@ -2,6 +2,7 @@ import type { SerializedStyles } from '@emotion/react';
 import { css } from '@emotion/react';
 import type { ComponentTopBar } from '../../styleD/build/typescript/component/TopBar';
 import { componentTopBar } from '../../styleD/build/typescript/component/TopBar';
+import { type Breakpoint, from, until } from '../../styleD/utils/semantic/mq';
 import type { DeepPartial, Prettify } from '../../util/types';
 
 export type TopBarTheme = Prettify<ComponentTopBar>;
@@ -12,24 +13,79 @@ export const topBarStyles = (theme: TopBarTheme): SerializedStyles => {
 	return css`
 		background-color: ${theme['background-color']};
 		height: ${theme.height};
-		border: ${theme.border};
+		border-top: ${theme['border-top']};
+		border-right: ${theme['border-right']};
+		border-bottom: ${theme['border-bottom']};
+		border-left: ${theme['border-left']};
 		display: ${theme.display};
 		justify-content: ${theme['justify-content']};
 	`;
 };
 
-export const topBarContainerLeftStyles = (
+export const topBarContainerStyles = (
 	theme: TopBarTheme,
+	{
+		collapseBelow,
+		showUntil,
+	}: { collapseBelow?: Breakpoint; showUntil?: Breakpoint } = {},
 ): SerializedStyles => {
 	return css`
 		display: ${theme.display};
+
+		${collapseBelow &&
+		css`
+			${until[collapseBelow]} {
+				display: none;
+			}
+		`}
+
+		${showUntil &&
+		css`
+			${from[showUntil]} {
+				display: none;
+			}
+		`}
 	`;
 };
 
-export const topBarContainerRightStyles = (
+export const topBarSpacerStyles = (theme: TopBarTheme): SerializedStyles => {
+	return css`
+		margin-left: ${theme.spacer['margin-left']};
+	`;
+};
+
+export const topBarCollapsedNavMenuButtonStyles = (
+	theme: TopBarTheme,
+	menuOpen?: boolean,
+): SerializedStyles => {
+	return css`
+		margin: ${theme.collapsedNavMenu.button.margin};
+		padding: ${theme.collapsedNavMenu.button.padding};
+		background: ${theme.collapsedNavMenu.button.background};
+		border: ${theme.collapsedNavMenu.button.border};
+		cursor: ${theme.collapsedNavMenu.button.cursor};
+
+		&[data-hovered] {
+			background-color: ${theme.collapsedNavMenu.button.hovered[
+				'background-color'
+			]};
+		}
+
+		${menuOpen &&
+		css`
+			background-color: ${theme.collapsedNavMenu.button.active[
+				'background-color'
+			]};
+		`}
+	`;
+};
+
+export const topBarCollapsedNavMenuPopoverStyles = (
 	theme: TopBarTheme,
 ): SerializedStyles => {
 	return css`
-		display: ${theme.display};
+		background-color: ${theme.collapsedNavMenu.popover['background-color']};
+		border-left: ${theme.collapsedNavMenu.popover.border};
+		border-right: ${theme.collapsedNavMenu.popover.border};
 	`;
 };