+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 # @guardian/stand
 
+## 0.0.31
+
+### Patch Changes
+
+- bcb7595: Expose collapsed menu button color and specific border sides from TopBar theme
+
+## 0.0.30
+
+### Patch Changes
+
+- a32c7c5: Extract a more generic Autocomplete for export and use in TagAutocomplete
+
+## 0.0.29
+
+### Patch Changes
+
+- 443d0cf: Fix TopBar responsive menu positioning
+
+## 0.0.28
+
+### Patch Changes
+
+- 4466180: Form Inputs no longer set `max-width` and instead width should be defined by parent element
+- 077ae9c: Add breakpoints and responsive TopBar
+- c68d8a0: add new semantic colours to ts and css exports
+
+      	'disabled-inverse'
+      	'link'
+      	'link-hover'
+      	'link-pressed'
+
 ## 0.0.27
 
 ### Patch Changes